Override lookup timeout
In case of timeouts, I was overriding the lookup time to 10 seconds to
make is visible on the histogram. That is not needed. The default
timeout is 2 seconds, which is already too high and does not need
special emphasis in histogram. You can also override default timeout
value from config.